zuoci-api-boot/src/main/java/com/easystudio/api/zuoci/service/PhraseService.java
package com.easystudio.api.zuoci.service;

import com.easystudio.api.zuoci.entity.Phrase;
import com.easystudio.api.zuoci.model.PhraseData;
import com.easystudio.api.zuoci.repository.PhraseRepository;
import com.easystudio.api.zuoci.translator.PhraseTranslator;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.stereotype.Service;

@Service
public class PhraseService {

    @Autowired
    private PhraseRepository repository;

    public void createPhrase(PhraseData data) {
        PhraseTranslator translator = new PhraseTranslator();
        Phrase phrase = translator.toPhraseEntity(data);
        repository.save(phrase);
    }

    public Page<Phrase> searchPhrase(boolean isValid, boolean isVisible, Pageable page) {
        Sort sort = new Sort(Sort.Direction.DESC, "createdTime");
        Pageable pageable = new PageRequest(page.getPageNumber(), page.getPageSize(), sort);
        return repository.findByIsValidAndIsVisible(isValid, isVisible, pageable);
    }
}
